+++ src/JK.Common/CHANGELOG.md
@@ -15,6 +15,7 @@ Unless otherwise noted, all changes made by [@jeremyknight-me](https://github.co
 - Added `Parse` methods to `StringExtensions`
 - Added `Range` and `Index` "polyfills"
 - Added Regex code generation to `RegexHelper`
+- Added `HasItem` extension methods for `ICollection<T>` and `IReadOnlyCollection<T>`
 
 ### Changed